mesosScheduler: Changes the task lookup implementation.
Currently, Gatekeeper sends a request to for the whole Mesos state
every time there is a token request. This can lead to request timeouts
if Mesos handles over 1000 tasks, since the response is over 4MB.
In order to solve this, I changed the getMesosTask implementation
to request data only for the task we are interested in. Additionally,
mesosScheduler now caches the framework names so as to not send
a request for framework info every time.